- добавлены коммиты и докстриги - создан файл settings.ini - создан файл utils.py - незначительно переработана структура программы - доработан README.md
