+++ README.md
@@ -1,8 +1,30 @@
->Курсовая 5. Работа с базами данных.
-> >1. Реализована работа с API HH.RU
-> >2. Создана БД при помощи SQL-запросов
-> >3. Реализован класс для работы с БД Postgres
-> 1. В файле api_hh_func.py хранятся функции для парса страниц работодателей.
-> 2. Файл queries.sql хранит запросы создания БД и таблиц.
-> 3. DBManager.py - класс для взаимодействия с БД Postgres. Запросы на заполнения таблиц и получение данных.
-> 4. main.py - файл запуска проекта. Перед его запуском следует выполнить запросы из файла queries.sql.
+<h1 align="center">Курсовая работа 5. Работа с базами данных.</h1>
+<h2>Описание</h2>
+<p>В данном проекте реализовано получение данных о компаниях и вакансиях с сайта hh.ru. Полученные данные загружаются в 
+БД PostgreSQL и фильтруются по заданным программой параметрам.</p>
+<h2>Установка</h2>
+<ol type="1">
+<li>Необходимо создать базу данных с таблицами по SQL-запросу из файла queries.sql при помощи терминала или pgAdmin 4.</li>
+<li>В файле settings.ini прописать свои параметры для подключения к БД:
+<p>   database=(название БД)</p>
+<p>   username=(имя пользователя (по умолчанию postgres))</p>
+<p>   password=(пароль доступа к серверу БД)</p></li>
+<li>По возможности измените id компаний, чьи вакансии хотите отследить, в файле main.py, функция create_db_tables(), список employers_id.</li>
+<li>Запустите main.py.</li>
+</ol>
+<h2>Работа ПО</h2>
+<ol type="1">
+<li>По запуску main.py заполняются таблицы БД.</li>
+<li>Выполняются SQL-запросы, после нажатия клавиши Enter.</li>
+<li>Результат запроса выводится на экран.</li>
+<a href="https://ibb.co/Lgck0DQ"><img src="https://i.ibb.co/GM8xdfs/projinterface.png" alt="projinterface" border="0"></a>
+</ol>
+<h2>Состав ПО</h2>
+<ul>
+<li>В файле api_hh_func.py хранятся функции для парса страниц работодателей.</li>
+<li>Файл queries.sql хранит запросы создания БД и таблиц.</li>
+<li>DBManager.py – класс для взаимодействия с БД Postgres. Запросы на заполнения таблиц и получение данных.</li>
+<li>main.py – файл запуска проекта. Перед его запуском следует выполнить запросы из файла queries.sql.</li>
+<li>settings.ini – индивидуальные параметры авторизации пользователя в БД PostgreSQL.</li>
+<li>utils.py – хранит функции необходимые для запуска проекта.</li>
+</ul>
