+++ settings.ini
@@ -1,0 +1,5 @@
+[DBConfig]
+host=localhost
+database=hh_vacancies
+username=postgres
+password=Mentos19
